fastapi_backend/requirements.txt

fastapi==0.111.0
uvicorn==0.30.1
SQLAlchemy==2.0.30
PyMySQL==1.1.1
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
